+++ R/utilities.R
@@ -1,3 +1,30 @@
+
+#' A helper function that tests whether an object is either NULL _or_
+#' a list of NULLs
+#'
+#' @keywords internal
+is.NullOb <- function(x) is.null(x) | all(sapply(x, is.null))
+
+#' Recursively step down into list, removing all such objects
+#'
+#' @keywords internal
+rmNullObs <- function(x) {
+  x <- Filter(Negate(is.NullOb), x)
+  lapply(x, function(x) if (is.list(x)) rmNullObs(x) else x)
+}
+
+# safe cbind that removes df with no rows
+my_cbind <- function(...){
+  dots <- list(...)
+
+  nrows <- vapply(dots, function(x) nrow(x) > 0, logical(1))
+
+  dots <- dots[nrows]
+
+  do.call(cbind, args = dots)
+
+}
+
 # purrr's map_df without dplyr
 my_map_df <- function(.x, .f, ...){