+++ NEWS.md
@@ -1,13 +1,17 @@
-# 0.1.1
+# 0.2.0
 
-## Changes 
+* Added an example Shiny app that calls the Speech API
+* Fixed bug where cbind of any missing API content raised an error (#28)
+* Add Google text to speech via `gl_talk()` (#39)
+* Add classify text endpoint for `gl_nlp()` (#20)
+
+# 0.1.1
 
 * Fix bug where `gl_speech()` only returned first few seconds of translation when asynch (#23)
 * CRAN version carries stable API, GitHub version for beta features
 
 # 0.1.0
 
-## First release
 
 * Natural language API via `gl_nlp`
 * Speech annotation via `gl_speech_recognise`